cv/requirements.txt
opencv-python
mediapipe
paho-mqtt
numpy
scipy
